Conflicts:
	cmds/bootanimation/Android.mk
	cmds/bootanimation/BootAnimation.cpp
	cmds/idmap/scan.cpp
	core/java/android/app/DownloadManager.java
	core/java/android/bluetooth/BluetoothHeadset.java
	core/java/android/provider/Settings.java
	core/java/android/widget/PopupWindow.java
	core/jni/android_hardware_Camera.cpp
	core/jni/android_media_AudioFormat.h
	core/jni/fd_utils-inl.h
	core/res/AndroidManifest.xml
	core/res/res/values-ar/strings.xml
	core/res/res/values-bn-rBD/strings.xml
	core/res/res/values-de/strings.xml
	core/res/res/values-es-rUS/strings.xml
	core/res/res/values-es/strings.xml
	core/res/res/values-fa/strings.xml
	core/res/res/values-fr/strings.xml
	core/res/res/values-gu-rIN/strings.xml
	core/res/res/values-hi/strings.xml
	core/res/res/values-in/strings.xml
	core/res/res/values-it/strings.xml
	core/res/res/values-km-rKH/strings.xml
	core/res/res/values-kn-rIN/strings.xml
	core/res/res/values-ko/strings.xml
	core/res/res/values-lo-rLA/strings.xml
	core/res/res/values-ml-rIN/strings.xml
	core/res/res/values-mr-rIN/strings.xml
	core/res/res/values-my-rMM/strings.xml
	core/res/res/values-pa-rIN/strings.xml
	core/res/res/values-pl/strings.xml
	core/res/res/values-pt/strings.xml
	core/res/res/values-ro/strings.xml
	core/res/res/values-ru/strings.xml
	core/res/res/values-sk/strings.xml
	core/res/res/values-sl/strings.xml
	core/res/res/values-sr/strings.xml
	core/res/res/values-sw/strings.xml
	core/res/res/values-ta-rIN/strings.xml
	core/res/res/values-te-rIN/strings.xml
	core/res/res/values-th/strings.xml
	core/res/res/values-tl/strings.xml
	core/res/res/values-tr/strings.xml
	core/res/res/values-uk/strings.xml
	core/res/res/values-ur-rPK/strings.xml
	core/res/res/values-vi/strings.xml
	core/res/res/values-zh-rCN/strings.xml
	core/res/res/values-zh-rHK/strings.xml
	core/res/res/values-zh-rTW/strings.xml
	core/res/res/values/config.xml
	core/res/res/values/strings.xml
	core/res/res/values/symbols.xml
	media/java/android/media/MediaFile.java
	media/java/android/media/SoundPool.java
	packages/Keyguard/src/com/android/keyguard/KeyguardUpdateMonitor.java
	packages/SettingsLib/src/com/android/settingslib/drawer/SettingsDrawerActivity.java
	packages/SystemUI/res/drawable/ic_qs_signal_4g_plus.xml
	packages/SystemUI/res/drawable/ic_qs_signal_lte_plus.xml
	packages/SystemUI/res/drawable/stat_sys_data_fully_connected_4g_plus.xml
	packages/SystemUI/res/drawable/stat_sys_data_fully_connected_lte_plus.xml
	packages/SystemUI/res/values/config.xml
	packages/SystemUI/res/values/strings.xml
	packages/SystemUI/src/com/android/systemui/statusbar/phone/KeyguardBottomAreaView.java
	packages/SystemUI/src/com/android/systemui/statusbar/phone/KeyguardStatusBarView.java
	packages/SystemUI/src/com/android/systemui/statusbar/phone/PhoneStatusBar.java
	packages/SystemUI/src/com/android/systemui/statusbar/policy/MobileSignalController.java
	packages/SystemUI/src/com/android/systemui/statusbar/policy/NetworkController.java
	packages/SystemUI/src/com/android/systemui/statusbar/policy/NetworkControllerImpl.java
	packages/SystemUI/src/com/android/systemui/statusbar/policy/TelephonyIcons.java
	services/core/java/com/android/server/AlarmManagerService.java
	services/core/java/com/android/server/ConnectivityService.java
	services/core/java/com/android/server/LockSettingsService.java
	services/core/java/com/android/server/NetworkManagementService.java
	services/core/java/com/android/server/connectivity/Tethering.java
	services/core/java/com/android/server/display/LocalDisplayAdapter.java
	services/core/java/com/android/server/location/GpsXtraDownloader.java
	services/core/java/com/android/server/notification/NotificationManagerService.java
	services/core/java/com/android/server/notification/ScheduleConditionProvider.java
	services/core/java/com/android/server/pm/PackageManagerService.java
	services/java/com/android/server/SystemServer.java
	telecomm/java/android/telecom/Conference.java
	telephony/java/android/telephony/CarrierConfigManager.java
	telephony/java/android/telephony/DisconnectCause.java
	telephony/java/android/telephony/RadioAccessFamily.java
	telephony/java/android/telephony/ServiceState.java
	telephony/java/com/android/internal/telephony/ITelephony.aidl
	telephony/java/com/android/internal/telephony/RILConstants.java

+++ services/voiceinteraction/java/com/android/server/voiceinteraction/DatabaseHelper.java
@@ -27,6 +27,9 @@ import android.hardware.soundtrigger.SoundTrigger.KeyphraseSoundModel;
 import android.text.TextUtils;
 import android.util.Slog;
 
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.List;
 import java.util.Locale;
 import java.util.UUID;
 
@@ -40,11 +43,12 @@ public class DatabaseHelper extends SQLiteOpenHelper {
     static final boolean DBG = false;
 
     private static final String NAME = "sound_model.db";
-    private static final int VERSION = 4;
+    private static final int VERSION = 6;
 
     public static interface SoundModelContract {
         public static final String TABLE = "sound_model";
         public static final String KEY_MODEL_UUID = "model_uuid";
+        public static final String KEY_VENDOR_UUID = "vendor_uuid";
         public static final String KEY_KEYPHRASE_ID = "keyphrase_id";
         public static final String KEY_TYPE = "type";
         public static final String KEY_DATA = "data";
@@ -57,14 +61,19 @@ public class DatabaseHelper extends SQLiteOpenHelper {
     // Table Create Statement
     private static final String CREATE_TABLE_SOUND_MODEL = "CREATE TABLE "
             + SoundModelContract.TABLE + "("
-            + SoundModelContract.KEY_MODEL_UUID + " TEXT PRIMARY KEY,"
+            + SoundModelContract.KEY_MODEL_UUID + " TEXT,"
+            + SoundModelContract.KEY_VENDOR_UUID + " TEXT,"
             + SoundModelContract.KEY_KEYPHRASE_ID + " INTEGER,"
             + SoundModelContract.KEY_TYPE + " INTEGER,"
             + SoundModelContract.KEY_DATA + " BLOB,"
             + SoundModelContract.KEY_RECOGNITION_MODES + " INTEGER,"
             + SoundModelContract.KEY_LOCALE + " TEXT,"
             + SoundModelContract.KEY_HINT_TEXT + " TEXT,"
-            + SoundModelContract.KEY_USERS + " TEXT" + ")";
+            + SoundModelContract.KEY_USERS + " TEXT,"
+            + "PRIMARY KEY (" + SoundModelContract.KEY_KEYPHRASE_ID + ","
+                              + SoundModelContract.KEY_LOCALE + ","
+                              + SoundModelContract.KEY_USERS + ")"
+            + ")";
 
     public DatabaseHelper(Context context) {
         super(context, NAME, null, VERSION);
@@ -78,9 +87,57 @@ public class DatabaseHelper extends SQLiteOpenHelper {
 
     @Override
     public void onUpgrade(SQLiteDatabase db, int oldVersion, int newVersion) {
-        // TODO: For now, drop older tables and recreate new ones.
-        db.execSQL("DROP TABLE IF EXISTS " + SoundModelContract.TABLE);
-        onCreate(db);
+        if (oldVersion < 4) {
+            // For old versions just drop the tables and recreate new ones.
+            db.execSQL("DROP TABLE IF EXISTS " + SoundModelContract.TABLE);
+            onCreate(db);
+        } else {
+            // In the jump to version 5, we added support for the vendor UUID.
+            if (oldVersion == 4) {
+                Slog.d(TAG, "Adding vendor UUID column");
+                db.execSQL("ALTER TABLE " + SoundModelContract.TABLE + " ADD COLUMN "
+                        + SoundModelContract.KEY_VENDOR_UUID + " TEXT");
+                oldVersion++;
+            }
+        }
+        if (oldVersion == 5) {
+            // We need to enforce the new primary key constraint that the
+            // keyphrase id, locale, and users are unique. We have to first pull
+            // everything out of the database, remove duplicates, create the new
+            // table, then push everything back in.
+            String selectQuery = "SELECT * FROM " + SoundModelContract.TABLE;
+            Cursor c = db.rawQuery(selectQuery, null);
+            List<SoundModelRecord> old_records = new ArrayList<SoundModelRecord>();
+            try {
+                if (c.moveToFirst()) {
+                    do {
+                        try {
+                            old_records.add(new SoundModelRecord(5, c));
+                        } catch (Exception e) {
+                            Slog.e(TAG, "Failed to extract V5 record", e);
+                        }
+                    } while (c.moveToNext());
+                }
+            } finally {
+                c.close();
+            }
+            db.execSQL("DROP TABLE IF EXISTS " + SoundModelContract.TABLE);
+            onCreate(db);
+            for (SoundModelRecord record : old_records) {
+                if (record.ifViolatesV6PrimaryKeyIsFirstOfAnyDuplicates(old_records)) {
+                    try {
+                        long return_value = record.writeToDatabase(6, db);
+                        if (return_value == -1) {
+                            Slog.e(TAG, "Database write failed " + record.modelUuid + ": "
+                                    + return_value);
+                        }
+                    } catch (Exception e) {
+                        Slog.e(TAG, "Failed to update V6 record " + record.modelUuid, e);
+                    }
+                }
+            }
+            oldVersion++;
+        }
     }
 
     /**
@@ -93,6 +150,9 @@ public class DatabaseHelper extends SQLiteOpenHelper {
             SQLiteDatabase db = getWritableDatabase();
             ContentValues values = new ContentValues();
             values.put(SoundModelContract.KEY_MODEL_UUID, soundModel.uuid.toString());
+            if (soundModel.vendorUuid != null) {
+                values.put(SoundModelContract.KEY_VENDOR_UUID, soundModel.vendorUuid.toString());
+            }
             values.put(SoundModelContract.KEY_TYPE, SoundTrigger.SoundModel.TYPE_KEYPHRASE);
             values.put(SoundModelContract.KEY_DATA, soundModel.data);
 
@@ -176,6 +236,11 @@ public class DatabaseHelper extends SQLiteOpenHelper {
                             continue;
                         }
 
+                        String vendorUuidString = null;
+                        int vendorUuidColumn = c.getColumnIndex(SoundModelContract.KEY_VENDOR_UUID);
+                        if (vendorUuidColumn != -1) {
+                            vendorUuidString = c.getString(vendorUuidColumn);
+                        }
                         byte[] data = c.getBlob(c.getColumnIndex(SoundModelContract.KEY_DATA));
                         int recognitionModes = c.getInt(
                                 c.getColumnIndex(SoundModelContract.KEY_RECOGNITION_MODES));
@@ -212,9 +277,12 @@ public class DatabaseHelper extends SQLiteOpenHelper {
                         Keyphrase[] keyphrases = new Keyphrase[1];
                         keyphrases[0] = new Keyphrase(
                                 keyphraseId, recognitionModes, modelLocale, text, users);
+                        UUID vendorUuid = null;
+                        if (vendorUuidString != null) {
+                            vendorUuid = UUID.fromString(vendorUuidString);
+                        }
                         KeyphraseSoundModel model = new KeyphraseSoundModel(
-                                UUID.fromString(modelUuid),
-                                null /* FIXME use vendor UUID */, data, keyphrases);
+                                UUID.fromString(modelUuid), vendorUuid, data, keyphrases);
                         if (DBG) {
                             Slog.d(TAG, "Found SoundModel for the given keyphrase/locale/user: "
                                     + model);
@@ -255,5 +323,94 @@ public class DatabaseHelper extends SQLiteOpenHelper {
             users[i] = Integer.parseInt(usersStr[i]);
         }
         return users;
+    }
+
+    private static class SoundModelRecord {
+        public final String modelUuid;
+        public final String vendorUuid;
+        public final int keyphraseId;
+        public final int type;
+        public final byte[] data;
+        public final int recognitionModes;
+        public final String locale;
+        public final String hintText;
+        public final String users;
+
+        public SoundModelRecord(int version, Cursor c) {
+            modelUuid = c.getString(c.getColumnIndex(SoundModelContract.KEY_MODEL_UUID));
+            if (version >= 5) {
+                vendorUuid = c.getString(c.getColumnIndex(SoundModelContract.KEY_VENDOR_UUID));
+            } else {
+                vendorUuid = null;
+            }
+            keyphraseId = c.getInt(c.getColumnIndex(SoundModelContract.KEY_KEYPHRASE_ID));
+            type = c.getInt(c.getColumnIndex(SoundModelContract.KEY_TYPE));
+            data = c.getBlob(c.getColumnIndex(SoundModelContract.KEY_DATA));
+            recognitionModes = c.getInt(c.getColumnIndex(SoundModelContract.KEY_RECOGNITION_MODES));
+            locale = c.getString(c.getColumnIndex(SoundModelContract.KEY_LOCALE));
+            hintText = c.getString(c.getColumnIndex(SoundModelContract.KEY_HINT_TEXT));
+            users = c.getString(c.getColumnIndex(SoundModelContract.KEY_USERS));
+        }
+
+        private boolean V6PrimaryKeyMatches(SoundModelRecord record) {
+          return keyphraseId == record.keyphraseId && stringComparisonHelper(locale, record.locale)
+              && stringComparisonHelper(users, record.users);
+        }
+
+        // Returns true if this record is a) the only record with the same V6 primary key, or b) the
+        // first record in the list of all records that have the same primary key and equal data.
+        // It will return false if a) there are any records that have the same primary key and
+        // different data, or b) there is a previous record in the list that has the same primary
+        // key and data.
+        // Note that 'this' object must be inside the list.
+        public boolean ifViolatesV6PrimaryKeyIsFirstOfAnyDuplicates(
+                List<SoundModelRecord> records) {
+            // First pass - check to see if all the records that have the same primary key have
+            // duplicated data.
+            for (SoundModelRecord record : records) {
+                if (this == record) {
+                    continue;
+                }
+                // If we have different/missing data with the same primary key, then we should drop
+                // everything.
+                if (this.V6PrimaryKeyMatches(record) && !Arrays.equals(data, record.data)) {
+                    return false;
+                }
+            }
+
+            // We only want to return true for the first duplicated model.
+            for (SoundModelRecord record : records) {
+                if (this.V6PrimaryKeyMatches(record)) {
+                    return this == record;
+                }
+            }
+            return true;
+        }
+
+        public long writeToDatabase(int version, SQLiteDatabase db) {
+            ContentValues values = new ContentValues();
+            values.put(SoundModelContract.KEY_MODEL_UUID, modelUuid);
+            if (version >= 5) {
+                values.put(SoundModelContract.KEY_VENDOR_UUID, vendorUuid);
+            }
+            values.put(SoundModelContract.KEY_KEYPHRASE_ID, keyphraseId);
+            values.put(SoundModelContract.KEY_TYPE, type);
+            values.put(SoundModelContract.KEY_DATA, data);
+            values.put(SoundModelContract.KEY_RECOGNITION_MODES, recognitionModes);
+            values.put(SoundModelContract.KEY_LOCALE, locale);
+            values.put(SoundModelContract.KEY_HINT_TEXT, hintText);
+            values.put(SoundModelContract.KEY_USERS, users);
+
+            return db.insertWithOnConflict(
+                       SoundModelContract.TABLE, null, values, SQLiteDatabase.CONFLICT_REPLACE);
+        }
+
+        // Helper for checking string equality - including the case when they are null.
+        static private boolean stringComparisonHelper(String a, String b) {
+          if (a != null) {
+            return a.equals(b);
+          }
+          return a == b;
+        }
     }
 }
